Starting lib folder migration
