+++ .gitignore
@@ -21,3 +21,5 @@ dist/*
 logs/*.log
 npm-debug.log*
 *.zip
+
+built/*
