+++ package.json
@@ -39,11 +39,11 @@
   "scripts": {
     "setup": "node -r esm built/src/setup.js",
     "theme:install": "./scripts/theme-install.sh",
-    "theme:export": "./scripts/theme-export.sh",    
+    "theme:export": "./scripts/theme-export.sh",
     "prettier": "prettier --write \"src/**/*.ts\" \"theme/**/*.js\"",
     "start": "node -r esm built/src/index.js",
     "tslint": "tslint -c tslint.json 'src/**/*.ts'"
-  },  
+  },
   "dependencies": {
     "aws-sdk": "^2.455.0",
     "bcrypt": "^3.0.4",
@@ -63,7 +63,7 @@
     "helmet": "^3.13.0",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^0.14.3",
-    "jsonwebtoken": "^8.3.0",    
+    "jsonwebtoken": "^8.3.4",
     "lodash": "^4.17.13",
     "lru-cache": "^4.1.3",
     "lscache": "^1.3.0",
@@ -84,15 +84,21 @@
     "ua-parser-js": "^0.7.18",
     "webpack": "^4.17.1",
     "webpack-cli": "^3.1.0",
-    "winston": "^3.0.0",
+    "winston": "^3.2.1",
     "workbox-webpack-plugin": "^3.4.1",
-    "ws": "^6.0.0"
+    "ws": "^6.0.3"
   },
   "engines": {
     "node": "^8.0.0",
     "npm": "^5.0.0"
   },
   "devDependencies": {
-    "@types/jsonwebtoken": "^8.3.4"
+    "@types/express": "^4.17.1",
+    "@types/jsonwebtoken": "^8.3.4",
+    "@types/mongodb": "^3.3.6",
+    "@types/nodemailer": "^6.2.1",
+    "@types/nodemailer-smtp-transport": "^2.7.4",
+    "@types/winston": "^2.4.4",
+    "@types/ws": "^6.0.3"
   }
 }
