+++ src/lib/mongo.ts
@@ -1,11 +1,11 @@
-import winston from 'winston';
-import url from 'url';
-import { MongoClient } from 'mongodb';
+import * as winston from 'winston';
+import * as url from 'url';
+import { MongoClient, Db } from 'mongodb';
 import settings from './settings';
 
 const mongodbConnection = settings.mongodbServerUrl;
 const mongoPathName = url.parse(mongodbConnection).pathname;
-const dbName = mongoPathName.substring(mongoPathName.lastIndexOf('/') + 1);
+const dbName = mongoPathName!.substring(mongoPathName!.lastIndexOf('/') + 1);
 
 const RECONNECT_INTERVAL = 1000;
 const CONNECT_OPTIONS = {
@@ -22,7 +22,7 @@ const onReconnect = () => {
 	winston.info('MongoDB reconnected');
 };
 
-export let db = null;
+export let db: Db | null = null;
 
 const connectWithRetry = () => {
 	MongoClient.connect(mongodbConnection, CONNECT_OPTIONS, (err, client) => {
