--- src/lib/settings.js
@@ -1,1 +1,0 @@
-export { default } from '../../config/server';
